+++ src/preset/interceptors/commonCancelRequest.js
@@ -23,7 +23,7 @@
  * set config.notCancel = true 时则不拦截
  * @param {FilterFunc} [abortFilter] default: url === url, method === method
  * @param {CommonCancelOption} [option]
- * @return {{install(interceptors:Interceptors):void}}
+ * @return {{requestInfoCache: (RequestInfo | null)[], install(interceptors:Interceptors):void}}
  */
 export default function commonCancelRequest(abortFilter, option) {
   /** @type {CommonCancelOption} */
@@ -38,11 +38,12 @@ export default function commonCancelRequest(abortFilter, option) {
   if (!abortFilter && typeof abortFilter !== 'function') {
     abortFilter = (currentConfig, storedConfig) => currentConfig.url === storedConfig.url && currentConfig.init.method === storedConfig.init.method;
   }
+  /** @type {(RequestInfo | null)[]} */
+  let requestInfoCache = [];
 
   return {
+    requestInfoCache,
     install(interceptors) {
-      /** @type {(RequestInfo | null)[]} */
-      let cacheArr = [];
       let requestId = 1;
 
       interceptors.request.use((url, init) => {
@@ -56,21 +57,30 @@ export default function commonCancelRequest(abortFilter, option) {
         init._commonCancelRequest = { requestId };
 
         if (window.AbortController) {
-          const abController = new AbortController();
-          init.signal = abController.signal;
+          /** @type {AbortController} */
+          let abController = init.abortController || new AbortController();
+          if (init.signal) {
+            // if user has set signal, use user's signal
+            // chrome 116 can use AbortSignal.any
+            if (AbortSignal.any) {
+              init.signal = AbortSignal.any([init.signal, abController.signal]);
+            }
+          } else {
+            init.signal = abController.signal;
+          }
           storedObj._controller = abController;
         } else {
           init._commonCancelRequest.canceled = false;
         }
         // check if has pending request
-        for (let i = 0; i < cacheArr.length; i++) {
-          const storedConfig = cacheArr[i];
+        for (let i = 0; i < requestInfoCache.length; i++) {
+          const storedConfig = requestInfoCache[i];
           if (!storedConfig) continue;
           if (abortFilter?.({ url, init, requestId }, storedConfig)) {
             if (window.AbortController) {
               hasPendingRequest = true; // not push to cache
               storedConfig._controller?.abort(`commonCancelRequest: cancel a request(${storedConfig.url})`); // abort request
-              cacheArr[i] = storedObj; // replace old cahce with new
+              requestInfoCache[i] = storedObj; // replace old cahce with new
             } else {
               // sign as cancelled，deal in response
               storedConfig.canceled = true;
@@ -80,24 +90,24 @@ export default function commonCancelRequest(abortFilter, option) {
 
         // if not has pending request, add to cache
         if (!hasPendingRequest) {
-          cacheArr.push(storedObj);
+          requestInfoCache.push(storedObj);
         }
         return init;
       });
 
       interceptors.response.use((data, { url, init }) => {
         let isReject = false;
-        for (let i = 0; i < cacheArr.length; i++) {
-          const item = cacheArr[i];
+        for (let i = 0; i < requestInfoCache.length; i++) {
+          const item = requestInfoCache[i];
           if (!item) continue;
           if (item.requestId === init._commonCancelRequest?.requestId) {
             isReject = Boolean(item?.canceled);
-            cacheArr[i] = null;
+            requestInfoCache[i] = null;
             break;
           }
         }
-        if (cacheArr.length > assignedOption.gcCacheArrNum) {
-          cacheArr = cacheArr.filter(Boolean); // 回收对象
+        if (requestInfoCache.length > assignedOption.gcCacheArrNum) {
+          requestInfoCache = requestInfoCache.filter(Boolean); // 回收对象
         }
         return isReject ? Promise.reject(`Request: ${url} has been ignored.`) : data;
       });